pyproject.toml
[tool.poetry]
name = "rr-bots"
version = "0.1.0"
description = ""
authors = ["JustLenard <justlenard.justme@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.18.1"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
